src/send-telegram/requirements.txt
fdk>=0.1.44
python-telegram-bot